refactor(service): extract PlaybackSnapshotManager and WidgetArtworkResolver from MusicService
Extract ~440 lines of widget artwork I/O and playback snapshot persistence
logic from MusicService into two focused, testable classes:

- WidgetArtworkResolver: handles artwork URI resolution, byte loading,
  caching, and failure retry for Glance home-screen widgets
- PlaybackSnapshotManager: manages capture, persistence, and debounced
  scheduling of the playback queue snapshot used for state restoration

MusicService now delegates to these components via lazy-initialized fields,
keeping only the orchestration glue (restorePlaybackQueueSnapshotIfNeeded)
that requires deep player access.
