+++ app/src/main/java/com/theveloper/pixelplay/data/service/widget/WidgetArtworkResolver.kt
@@ -1,0 +1,332 @@
+package com.theveloper.pixelplay.data.service.widget
+
+import android.content.Context
+import android.graphics.Bitmap
+import android.net.Uri
+import android.os.SystemClock
+import androidx.media3.common.MediaMetadata
+import androidx.media3.common.util.UnstableApi
+import coil.imageLoader
+import coil.request.CachePolicy
+import coil.request.ImageRequest
+import coil.size.Precision
+import com.theveloper.pixelplay.data.model.Song
+import com.theveloper.pixelplay.data.repository.MusicRepository
+import com.theveloper.pixelplay.data.service.player.DualPlayerEngine
+import com.theveloper.pixelplay.utils.AlbumArtUtils
+import com.theveloper.pixelplay.utils.ArtworkTransportSanitizer
+import com.theveloper.pixelplay.utils.LocalArtworkUri
+import com.theveloper.pixelplay.utils.MediaItemBuilder
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.withContext
+import timber.log.Timber
+import java.io.ByteArrayOutputStream
+import java.io.InputStream
+import java.net.HttpURLConnection
+import java.net.URL
+
+/**
+ * Resolves, caches, and loads artwork bytes for Glance home-screen widgets.
+ *
+ * Extracted from [MusicService] to isolate widget artwork I/O concerns
+ * (URI resolution, byte loading, failure caching) from the service's
+ * media-session lifecycle management.
+ */
+@UnstableApi
+class WidgetArtworkResolver(
+    private val context: Context,
+    private val engine: DualPlayerEngine,
+    private val musicRepository: MusicRepository,
+) {
+
+    private var cachedWidgetArtSourceKey: String? = null
+    private var cachedWidgetArtResolvedUri: String? = null
+    private var cachedWidgetArtBytes: ByteArray? = null
+    private var cachedWidgetArtLoadFailureKey: String? = null
+    private var cachedWidgetArtLoadFailureAtMs: Long = 0L
+
+    fun invalidateCachedWidgetArtwork() {
+        cachedWidgetArtSourceKey = null
+        cachedWidgetArtResolvedUri = null
+        cachedWidgetArtBytes = null
+        cachedWidgetArtLoadFailureKey = null
+        cachedWidgetArtLoadFailureAtMs = 0L
+    }
+
+    fun isCurrentWidgetArtworkBackedByTelegram(): Boolean {
+        val currentItem = engine.masterPlayer.currentMediaItem ?: return false
+        val metadata = currentItem.mediaMetadata
+        val contentUriString = currentItem.localConfiguration?.uri?.toString()
+            ?: metadata.extras?.getString(MediaItemBuilder.EXTERNAL_EXTRA_CONTENT_URI)
+        val artworkUriString = resolveStoredArtworkUriString(metadata)
+        return contentUriString?.startsWith("telegram://") == true ||
+            artworkUriString?.startsWith("telegram_art://") == true
+    }
+
+    suspend fun getAlbumArtForWidget(
+        mediaId: String?,
+        embeddedArt: ByteArray?,
+        artUris: List<Uri>,
+    ): Pair<ByteArray?, String?> = withContext(Dispatchers.IO) {
+        val sanitizedFromEmbedded = embeddedArt?.takeIf { it.isNotEmpty() }?.let { bytes ->
+            runCatching {
+                ArtworkTransportSanitizer.sanitizeEncodedBytes(
+                    data = bytes,
+                    config = ArtworkTransportSanitizer.WIDGET_CONFIG,
+                )
+            }.getOrNull()
+        }
+        val candidateUriStrings = LinkedHashSet<String>().apply {
+            artUris.forEach { candidate ->
+                candidate.toString()
+                    .takeIf { it.isNotBlank() }
+                    ?.let(::add)
+            }
+        }.toList()
+        val preferredUriString = candidateUriStrings.firstOrNull()
+        val sourceKey = buildWidgetArtworkSourceKey(
+            mediaId = mediaId,
+            candidateUriStrings = candidateUriStrings,
+        )
+
+        if (sanitizedFromEmbedded != null) {
+            cachedWidgetArtSourceKey = sourceKey
+            cachedWidgetArtResolvedUri = preferredUriString
+            cachedWidgetArtBytes = sanitizedFromEmbedded
+            cachedWidgetArtLoadFailureKey = null
+            cachedWidgetArtLoadFailureAtMs = 0L
+            return@withContext sanitizedFromEmbedded to preferredUriString
+        }
+
+        if (sourceKey != null && sourceKey == cachedWidgetArtSourceKey && cachedWidgetArtBytes != null) {
+            return@withContext cachedWidgetArtBytes to (cachedWidgetArtResolvedUri ?: preferredUriString)
+        }
+        if (sourceKey != null && sourceKey == cachedWidgetArtLoadFailureKey) {
+            val failureAgeMs = SystemClock.elapsedRealtime() - cachedWidgetArtLoadFailureAtMs
+            if (failureAgeMs < WIDGET_ART_FAILURE_RETRY_MS) {
+                return@withContext null to preferredUriString
+            }
+        }
+
+        val repositoryArtUriString = if (mediaId.isNullOrBlank()) {
+            null
+        } else {
+            resolveRepositoryArtworkUri(mediaId)?.toString()
+        }
+        val resolvedUriStrings = LinkedHashSet<String>().apply {
+            addAll(candidateUriStrings)
+            repositoryArtUriString
+                ?.takeIf { it.isNotBlank() }
+                ?.let(::add)
+        }
+
+        for (candidateUriString in resolvedUriStrings) {
+            val candidateUri = parseArtworkUriString(candidateUriString) ?: continue
+            val loadedBytes = loadArtworkBytes(candidateUri)
+            if (loadedBytes != null) {
+                cachedWidgetArtSourceKey = sourceKey
+                cachedWidgetArtResolvedUri = candidateUriString
+                cachedWidgetArtBytes = loadedBytes
+                cachedWidgetArtLoadFailureKey = null
+                cachedWidgetArtLoadFailureAtMs = 0L
+                return@withContext loadedBytes to candidateUriString
+            }
+        }
+
+        cachedWidgetArtLoadFailureKey = sourceKey
+        cachedWidgetArtLoadFailureAtMs = SystemClock.elapsedRealtime()
+        return@withContext null to (repositoryArtUriString ?: preferredUriString)
+    }
+
+    fun resolveStoredArtworkUriString(metadata: MediaMetadata?): String? {
+        metadata ?: return null
+        return metadata.extras
+            ?.getString(MediaItemBuilder.EXTERNAL_EXTRA_ALBUM_ART)
+            ?.takeIf { it.isNotBlank() }
+            ?: metadata.artworkUri
+                ?.toString()
+                ?.takeIf { it.isNotBlank() }
+    }
+
+    fun resolveWidgetArtworkUriCandidates(
+        metadata: MediaMetadata?,
+        preferredArtworkUri: Uri? = null,
+    ): List<Uri> {
+        val candidates = LinkedHashSet<String>()
+        preferredArtworkUri
+            ?.toString()
+            ?.takeIf { it.isNotBlank() }
+            ?.let(candidates::add)
+        resolveStoredArtworkUriString(metadata)?.let(candidates::add)
+        metadata?.artworkUri
+            ?.toString()
+            ?.takeIf { it.isNotBlank() }
+            ?.let(candidates::add)
+        return candidates.mapNotNull(::parseArtworkUriString)
+    }
+
+    fun resolveArtworkUri(metadata: MediaMetadata?): Uri? {
+        metadata ?: return null
+        metadata.artworkUri?.let { return it }
+        val extrasUri = metadata.extras
+            ?.getString(MediaItemBuilder.EXTERNAL_EXTRA_ALBUM_ART)
+            ?.takeIf { it.isNotBlank() }
+            ?: return null
+        return parseArtworkUriString(extrasUri)
+    }
+
+    suspend fun loadArtworkBytes(uri: Uri): ByteArray? {
+        val uriString = uri.toString()
+        val scheme = uri.scheme?.lowercase()
+        val isLocalArtworkUri = LocalArtworkUri.isLocalArtworkUri(uriString)
+        return when {
+            isLocalArtworkUri || scheme == "content" || scheme == "file" || scheme == "android.resource" -> {
+                runCatching {
+                    AlbumArtUtils.openArtworkInputStream(context, uri)?.use { input ->
+                        readBytesCapped(input, ArtworkTransportSanitizer.WIDGET_CONFIG.sourceBytesLimit)
+                            ?.let { bytes ->
+                                ArtworkTransportSanitizer.sanitizeEncodedBytes(
+                                    data = bytes,
+                                    config = ArtworkTransportSanitizer.WIDGET_CONFIG,
+                                )
+                            }
+                    }
+                }.getOrElse { error ->
+                    Timber.tag(TAG).w(error, "Widget artwork read failed for local uri=%s", uri)
+                    null
+                }
+            }
+            scheme == "http" || scheme == "https" -> {
+                var connection: HttpURLConnection? = null
+                try {
+                    connection = (URL(uriString).openConnection() as? HttpURLConnection)
+                        ?: return null
+                    connection.connectTimeout = 4_000
+                    connection.readTimeout = 6_000
+                    connection.instanceFollowRedirects = true
+                    connection.doInput = true
+                    connection.inputStream.use { input ->
+                        readBytesCapped(input, ArtworkTransportSanitizer.WIDGET_CONFIG.sourceBytesLimit)
+                            ?.let { bytes ->
+                                ArtworkTransportSanitizer.sanitizeEncodedBytes(
+                                    data = bytes,
+                                    config = ArtworkTransportSanitizer.WIDGET_CONFIG,
+                                )
+                            }
+                    }
+                } catch (error: Exception) {
+                    Timber.tag(TAG).w(error, "Widget artwork read failed for remote uri=%s", uri)
+                    null
+                } finally {
+                    connection?.disconnect()
+                }
+            }
+            else -> loadArtworkBytesViaCoil(context, uri)
+        }
+    }
+
+    private fun parseArtworkUriString(rawArtworkUri: String?): Uri? {
+        if (rawArtworkUri.isNullOrBlank()) {
+            return null
+        }
+
+        return MediaItemBuilder.artworkUri(rawArtworkUri)
+            ?: if (rawArtworkUri.startsWith("/")) {
+                Uri.fromFile(java.io.File(rawArtworkUri))
+            } else {
+                runCatching { Uri.parse(rawArtworkUri) }.getOrNull()
+            }
+    }
+
+    private fun buildWidgetArtworkSourceKey(
+        mediaId: String?,
+        candidateUriStrings: List<String>,
+    ): String? {
+        val normalizedMediaId = mediaId?.takeIf { it.isNotBlank() }
+        if (normalizedMediaId == null && candidateUriStrings.isEmpty()) {
+            return null
+        }
+        return buildString {
+            normalizedMediaId?.let {
+                append("mediaId=")
+                append(it)
+            }
+            if (candidateUriStrings.isNotEmpty()) {
+                if (isNotEmpty()) append('|')
+                append(candidateUriStrings.joinToString(separator = ","))
+            }
+        }
+    }
+
+    suspend fun resolveRepositoryArtworkUri(mediaId: String?): Uri? {
+        val songId = mediaId?.takeIf { it.isNotBlank() } ?: return null
+        val song = withContext(Dispatchers.IO) {
+            musicRepository.getSong(songId).first()
+        } ?: return null
+
+        return MediaItemBuilder.artworkUri(song.albumArtUriString)
+            ?: song.albumArtUriString
+                ?.takeIf { it.isNotBlank() }
+                ?.let { raw ->
+                    if (raw.startsWith("/")) Uri.fromFile(java.io.File(raw))
+                    else runCatching { Uri.parse(raw) }.getOrNull()
+                }
+    }
+
+    private fun readBytesCapped(input: InputStream, maxBytes: Int): ByteArray? {
+        val output = ByteArrayOutputStream(DEFAULT_STREAM_BUFFER_SIZE * 4)
+        val buffer = ByteArray(DEFAULT_STREAM_BUFFER_SIZE)
+        var totalRead = 0
+        while (true) {
+            val read = input.read(buffer)
+            if (read <= 0) break
+            totalRead += read
+            if (totalRead > maxBytes) return null
+            output.write(buffer, 0, read)
+        }
+        return output.toByteArray().takeIf { it.isNotEmpty() }
+    }
+
+    companion object {
+        private const val TAG = "WidgetArtworkResolver"
+        private const val DEFAULT_STREAM_BUFFER_SIZE = 8 * 1024
+        private const val WIDGET_ART_FAILURE_RETRY_MS = 30_000L
+    }
+}
+
+suspend fun loadArtworkBytesViaCoil(context: Context, uri: Uri): ByteArray? {
+    val appContext = context.applicationContext
+    val request = ImageRequest.Builder(appContext)
+        .data(uri)
+        .size(
+            ArtworkTransportSanitizer.WIDGET_CONFIG.maxDimensionPx,
+            ArtworkTransportSanitizer.WIDGET_CONFIG.maxDimensionPx,
+        )
+        .precision(Precision.INEXACT)
+        .allowHardware(false)
+        .memoryCachePolicy(CachePolicy.ENABLED)
+        .networkCachePolicy(CachePolicy.ENABLED)
+        .build()
+
+    return runCatching {
+        val drawable = appContext.imageLoader.execute(request).drawable ?: return@runCatching null
+        val fallbackSizePx = ArtworkTransportSanitizer.WIDGET_CONFIG.maxDimensionPx
+        val bitmap = drawable.toBitmap(
+            width = drawable.intrinsicWidth.takeIf { it > 0 } ?: fallbackSizePx,
+            height = drawable.intrinsicHeight.takeIf { it > 0 } ?: fallbackSizePx,
+            config = Bitmap.Config.ARGB_8888,
+        )
+        val encodedBytes = ByteArrayOutputStream().use { output ->
+            bitmap.compress(Bitmap.CompressFormat.JPEG, 92, output)
+            output.toByteArray()
+        }
+        ArtworkTransportSanitizer.sanitizeEncodedBytes(
+            data = encodedBytes,
+            config = ArtworkTransportSanitizer.WIDGET_CONFIG,
+        )
+    }.getOrElse { error ->
+        Timber.tag("WidgetArtworkResolver").w(error, "Artwork read failed via Coil for uri=%s", uri)
+        null
+    }
+}
